# openapi-draft.yaml
openapi: 3.1.0
info:
  title: Chiijie Outfit Enterprise E-Commerce API
  version: 1.0.0
  description: Draft OpenAPI contract for identity, catalog, inventory, order, payment, and notification services.
servers:
  - url: https://api.chiijie-outfit.com/v1
    description: Production API endpoint
  - url: https://staging-api.chiijie-outfit.com/v1
    description: Staging API endpoint
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - code
        - message

    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - fullName
        - role
        - status

    ProductSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        slug:
          type: string
        priceCents:
          type: integer
        salePriceCents:
          type: integer
        status:
          type: string
        metadata:
          type: object
      required:
        - id
        - sku
        - name
        - slug
        - priceCents
        - status

    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          properties:
            description:
              type: string
            categoryId:
              type: string
              format: uuid
            brand:
              type: string
            attributes:
              type: object
            variants:
              type: array
              items:
                $ref: '#/components/schemas/ProductVariant'

    ProductVariant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        sku:
          type: string
        variantName:
          type: string
        priceCents:
          type: integer
        salePriceCents:
          type: integer
        availableStock:
          type: integer
        reservedStock:
          type: integer
        attributes:
          type: object
      required:
        - id
        - productId
        - sku
        - variantName
        - priceCents

    OrderItem:
      type: object
      properties:
        productVariantId:
          type: string
          format: uuid
        quantity:
          type: integer
        unitPriceCents:
          type: integer
        totalPriceCents:
          type: integer
      required:
        - productVariantId
        - quantity
        - unitPriceCents
        - totalPriceCents

    OrderSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderNumber:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        totalAmountCents:
          type: integer
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - status
        - paymentStatus
        - totalAmountCents
        - currency

    PaymentCreateRequest:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        idempotencyKey:
          type: string
      required:
        - orderId
        - userId
        - idempotencyKey

    PaymentCreateResponse:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        status:
          type: string
      required:
        - clientSecret
        - paymentIntentId
        - status

    WebhookEventResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

paths:
  /auth/login:
    post:
      summary: Login user and return access token
      tags:
        - Identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Authenticated response
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products:
    get:
      summary: List products
      tags:
        - Catalog
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductSummary'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer

  /products/{productId}:
    get:
      summary: Get product detail
      tags:
        - Catalog
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /inventory/{variantId}:
    get:
      summary: Retrieve inventory snapshot
      tags:
        - Inventory
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inventory detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  productVariantId:
                    type: string
                    format: uuid
                  availableStock:
                    type: integer
                  reservedStock:
                    type: integer
                  incomingStock:
                    type: integer
        '404':
          description: Inventory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders:
    post:
      summary: Create a new order
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                shippingAddress:
                  type: object
                billingAddress:
                  type: object
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderItem'
              required:
                - userId
                - items
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummary'

  /orders/{orderId}:
    get:
      summary: Retrieve order detail
      tags:
        - Order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OrderSummary'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrderItem'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/create:
    post:
      summary: Create a payment intent
      tags:
        - Payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCreateResponse'

  /payments/webhook:
    post:
      summary: Payment gateway webhook endpoint
      tags:
        - Payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEventResponse'
        '400':
          description: Invalid webhook event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /notifications/email:
    post:
      summary: Send transactional email
      tags:
        - Notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                  format: email
                subject:
                  type: string
                body:
                  type: string
              required:
                - to
                - subject
                - body
      responses:
        '202':
          description: Email request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string